Change the span@class=brl:computer into an element
We would really like to move to a scheme where we can have all the brl
elements inside spans. However for consistency sake I guess we should
move all elements at the same time. So this means that for the moment
and to be consistent we will create an element named brl:computer for
computer braille.

Hopefully in the future we can move most of the brl:* elements to spans.