Move the definition of TABLE_BASE_URI to xproc
so that the tests use the pipeline tables and the run-time (using xproc)
will be using the locally installed tables (including the generated
white lists).

In the future we will hopefully fix this so that both the tests and
run-time use the same method of finding the tables, i.e. tables from
pipeline module and white lists from file system. For that however we
need a newer liblouis module and probably have to wait for the next
pipeline release.