Change the API of the xproc to use dashes in params
as opposed to underscore. It seems strange to do this because
Daisyproducer uses underscore, then changes it to dashes to pass to the
pipeline module which changes it back to underscore to hand it off to
the xsl. But for consistency sake with the rest of the pipeline
modules (which use dashes in param names) we do this.