Add a Dockerfile to containerize the utfx testing
For better or worse for the old legacy dtbook2sbsform code we use utfx
for testing. Since utfx has been abandoned a long time ago it only
runs with antique versions of Java. So we create a Docker image that
contains all the old versions of all that utfx needs and with the help
of that we can test dtbook2sbsform again.